Added database seeder, and fixed migration/model

+++ database/migrations/2020_06_10_011444_create_broadcasts_table.php
@@ -17,8 +17,8 @@ class CreateBroadcastsTable extends Migration
             $table->id();
             $table->integer('freq');
             $table->string('station');
-            $table->tinyInteger('start');
-            $table->tinyInteger('end');
+            $table->integer('start');
+            $table->integer('end');
             $table->integer('days');
             $table->string('language');
             $table->decimal('power')->nullable();
